// shared/src/types.d.ts
export type AssetCategoryType = 'cash' | 'gold' | 'silver' | 'business' | 'property' | 'stocks' | 'crypto' | 'debts' | 'expenses';
export interface Asset {
    assetId: string;
    name: string;
    category: AssetCategoryType;
    subCategory: string;
    value: number;
    currency: string;
    description?: string;
    zakatEligible: boolean;
    createdAt: string;
    updatedAt: string;
}